+++ Catpocalypse/Assets/Scenes/Level1/NavMesh-NavMesh Surface 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 897d2b3f43a1c9944a800f7e6be1c520
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
