--- Catpocalypse/Assets/Scripts/CutScenes/ComicPage.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6479d7679918dfd4699a78de678017e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
