--- Catpocalypse/Assets/Scripts/Level Select.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 472f9dee85a30214288e0c9f1edc5afb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
