--- Catpocalypse/Assets/Scripts/Level Select/LevelSelectButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6659ec2f36693b841a037761bdd25345
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
