+++ Catpocalypse/Assets/Scripts/UI/LevelSelectUI/ClickableLevelTile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cae4bba34f5e84d4684c69211b6ae49d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
